desaparecer botones de acción para el método de pago predeterminado

+++ resources/views/livewire/payment-method.blade.php
@@ -72,14 +72,19 @@
                             </p>
                         </div>
 
-                        <div class="flex space-x-4">
-                            <button class="disabled:opacity-25" wire:click="defaultPaymentMethod('{{$paymentMethod->id}}')" wire:target="defaultPaymentMethod('{{$paymentMethod->id}}')" wire:loading.attr="disabled">
-                                <i class="fa-regular fa-star"></i>
-                            </button>
-                            <button class="disabled:opacity-25" wire:click="deletePaymentMethod('{{$paymentMethod->id}}')" wire:target="deletePaymentMethod('{{$paymentMethod->id}}')" wire:loading.attr="disabled">
-                                <i class="fa-regular fa-trash-can"></i>
-                            </button>
-                        </div>
+                        @if ($this->defaultPaymentMethod->id != $paymentMethod->id)
+                            
+                            <div class="flex space-x-4">
+                                <button class="disabled:opacity-25" wire:click="defaultPaymentMethod('{{$paymentMethod->id}}')" wire:target="defaultPaymentMethod('{{$paymentMethod->id}}')" wire:loading.attr="disabled">
+                                    <i class="fa-regular fa-star"></i>
+                                </button>
+                                <button class="disabled:opacity-25" wire:click="deletePaymentMethod('{{$paymentMethod->id}}')" wire:target="deletePaymentMethod('{{$paymentMethod->id}}')" wire:loading.attr="disabled">
+                                    <i class="fa-regular fa-trash-can"></i>
+                                </button>
+                            </div>
+
+                        @endif
+                           
 
                     </li>
                 @endforeach
